backend/requirements.txt
# Core API
fastapi
uvicorn
python-multipart

# Configuration & Security
python-dotenv
azure-identity
pydantic
pydantic-settings
PyJWT[crypto]
requests

# Web Scraping
beautifulsoup4
trafilatura
lxml_html_clean

# Database
pyodbc

# LLM & Tokenization
openai
tiktoken

# Document loading & splitting
pdfplumber               
pypdf                    
python-docx
langchain-text-splitters

# Vector database / Search
azure-search-documents

# Document storage
azure-storage-blob